backend/requirements.txt
fastapi
uvicorn
PyJWT
python-dotenv
httpx
requests
sqlalchemy
psycopg2-binary
python-multipart
passlib
bcrypt
pydantic
